Improve search result filtering
Searching for both the author and the title at the same time sometimes
results in no books being found. In addition, searches often return
irrelevant results instead of or in addition to actually desirable
results.

To solve both of these issues, the search functionality now first
searches for just the title and then filters the results by matching all
of the title and author tokens against them. If that leaves no results,
which may happen with very generic titles that generate many irrelevant
results, it will search for both the author and the title as before and
then apply filtering as well.